# [WIP] Reduce number of loop computations