--- a/distributed/client.py
+++ b/distributed/client.py
@@ -2517,7 +2517,6 @@ class Client(Node):
                 extra_keys.add(tokey(future.key))
             dsk2 = str_graph(dsk2_dict, extra_keys)
             dsk3 = {k: v for k, v in dsk2.items() if k is not v}
-            
 
             if restrictions:
                 restrictions = keymap(tokey, restrictions)