This PR tries to reduce the number of times we loop on the same objects and also tries to squeeze into as few loops as possible. This is a first pass of the chanes, hence this is still WIP.