@mrocklin 2020-05-01
Thank you for working on this @galipremsagar .  At some point it will be useful to see if this has a significant impact on timing in real-world applications.