@kkraus14 2020-05-02
@galipremsagar it would be good to test with the example here https://github.com/dask/dask/issues/6163#issuecomment-622571290 and post results to see if this is making any meaningful improvements.