@galipremsagar 2020-05-02
> @galipremsagar it would be good to test with the example here [dask/dask#6163 (comment)](https://github.com/dask/dask/issues/6163#issuecomment-622571290) and post results to see if this is making any meaningful improvements.

For the sample you pointed to the improvement in performance is about 0.3s. But as the number of records is increased we see 1s improvement.

```python
from distributed import Client
from dask.datasets import timeseries
from dask.dataframe.shuffle import shuffle


client = Client()

ddf_d = timeseries(start='2000-01-01', end='2015-01-01', partition_freq='1d')
ddf_d_2 = shuffle(ddf_d, "id", shuffle="tasks")

%time ddf_d_2 = ddf_d_2.persist()
```

**Master**
CPU times: user 45.5 s, sys: 1.07 s, total: 46.6 s
Wall time: 46.2 s

**This PR**
CPU times: user 44.2 s, sys: 1.46 s, total: 45.6 s
Wall time: 45.2 s