@mrocklin 2020-05-26
> How many times have you run this? ~1s could just be noise smile

This would be good to know

It looks like things have stalled here.  Is this still a priority?