@galipremsagar 2020-05-26
> > How many times have you run this? ~1s could just be noise smile
> 
> This would be good to know
> 
> It looks like things have stalled here. Is this still a priority?

@mrocklin @kkraus14 

From the benchmark numbers it looks more or less the same against master:

```python
In [1]: from distributed import Client 
   ...: from dask.datasets import timeseries 
   ...: from dask.dataframe.shuffle import shuffle 
   ...:  
   ...:  
   ...: client = Client() 
   ...:  
   ...: ddf_d = timeseries(start='2000-01-01', end='2015-01-01', partition_freq='1d') 
   ...: ddf_d_2 = shuffle(ddf_d, "id", shuffle="tasks")   
```

Master
```python
In [2]: %timeit ddf_d_2.persist()                                                                   
34.8 s ± 2.61 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

This PR
```python
In [2]: %timeit ddf_d_2.persist()                                                                   
34.5 s ± 2.72 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```