@quasiben 2020-05-27
@galipremsagar does this PR reduce temporary memory usage ? If not, would you consider these style changes then given similar performance ?