@mrocklin 2020-05-27
It sounds like we're looking for a reason to continue this work. It's also totally ok to not continue it, and focus efforts elsewhere.  My guess is that there are plenty of things to work on when it comes to accelerating the scheduler.