@galipremsagar 2020-05-27
It reduces memory usages, here is a memory profile. Not a significant reduction but similar noticeable difference across runs.

This PR
```
In [2]: %memit ddf_d_2.persist()                                                              
peak memory: 1383.89 MiB, increment: 1029.93 MiB
```
Master
```
In [2]: %memit ddf_d_2.persist()                                                              
peak memory: 1438.21 MiB, increment: 1083.59 MiB
```