@quasiben 2020-05-27
> It sounds like we're looking for a reason to continue this work. It's also totally ok to not continue it, and focus efforts elsewhere. My guess is that there are plenty of things to work on when it comes to accelerating the scheduler.

Agreed.  I don't think we should push on this PR anymore.  There is some gain but not a whole lot.  Due to the large number of changes perhaps we should abandon. 

If others feel strongly that this should be merged in they should feel free to comment.