@galipremsagar 2020-05-28
Given that these are micro-optimizations, here are both ways to it.

We can abandon this PR as the speed-ups are not noticeable and changes are more.
We can merge this PR based on the minor reduction in memory foot-print reported above.