@jacobtomlinson 2020-05-29
Just to weigh in here I think many of the changes are style changes, some make things more readable, but others make things less readable. I would actually argue that small increases in contributor/maintainer performance with readable code is more important than small increases in code performance.

However, a ~4% reduction in memory usage is not insignificant. (Also TIL about `%memit` so thanks!).

Therefore I would probably go either way with this. And generally in OSS maintenance if I'm on the fence about a change like this I tend to merge it, mainly for the community benefit of getting another developer with changes in the project.